商品SPU,SKU实现

+++ lyn-goods/goods-service/src/main/java/com/lyn/goods/service/mapper/GoodsInfoMapper.java
@@ -10,7 +10,7 @@ public interface GoodsInfoMapper {
 
     int deleteByExample(GoodsInfoExample example);
 
-    int deleteByPrimaryKey(Integer goodsId);
+    int deleteByPrimaryKey(Long goodsId);
 
     int insert(GoodsInfo record);
 
@@ -18,7 +18,7 @@ public interface GoodsInfoMapper {
 
     List<GoodsInfo> selectByExample(GoodsInfoExample example);
 
-    GoodsInfo selectByPrimaryKey(Integer goodsId);
+    GoodsInfo selectByPrimaryKey(Long goodsId);
 
     int updateByExampleSelective(@Param("record") GoodsInfo record, @Param("example") GoodsInfoExample example);
